requirements.txt
aioredis==2.0.1
autopep8==1.6.0
bcrypt==3.2.2
black==22.3.0
build==0.8.0
dotmap==1.3.30
environs==9.5.0
fastapi==0.79.0
httptools==0.4.0
install==1.3.5
ipykernel==6.13.0
loguru==0.6.0
passlib==1.7.4
pipdeptree==2.2.1
pipreqs==0.4.11
psycopg2==2.9.3
pydeps==1.10.18
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
rich==12.4.4
SQLAlchemy-Utils==0.38.3
sqlmodel==0.0.6
uvicorn==0.18.2
watchgod==0.8.2
